# backend/openapi.yaml
openapi: 3.1.0
info:
  title: CofradeNet API
  description: |
    API REST de la plataforma CofradeNet. Gestiona hermandades, bandas, procesiones,
    usuarios cofrades y el panel de administración de la plataforma.

    ## Autenticación
    La mayoría de los endpoints de escritura requieren un token JWT Bearer.
    Obtén el token con `POST /auth/login` e inclúyelo en la cabecera:
    `Authorization: Bearer <token>`

    ## Roles disponibles
    | Rol | Descripción |
    |-----|-------------|
    | `cofrade` | Usuario registrado. Puede seguir entidades, crear publicaciones y gestionar su perfil. |
    | `hermandad` | Administra el perfil de su hermandad y sus procesiones. |
    | `banda` | Administra el perfil de su banda y su agenda de actuaciones. |
    | `admin` | Acceso completo a la plataforma. |
  version: 1.0.0
  contact:
    name: Equipo CofradeNet
servers:
  - url: http://localhost:3000
    description: Desarrollo local
...
paths:
...
  /bandas/ciudad/{id}:
    get:
      tags: [bandas]
      summary: Listar bandas de una ciudad
      operationId: bandasPorCiudad
      security: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Bandas de la ciudad indicada.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Banda'
...
  /ciudades:
    get:
      tags: [ciudades]
      summary: Listar ciudades (solo ADMIN)
      operationId: listarCiudades
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de ciudades.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ciudad'
    post:
      tags: [ciudades]
      summary: Crear ciudad (solo ADMIN)
      operationId: crearCiudad
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCiudadRequest'
      responses:
        '201':
          description: Ciudad creada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ciudad'
  /ciudades/buscar:
    get:
      tags: [ciudades]
      summary: Buscar ciudades por nombre (solo ADMIN)
      operationId: buscarCiudad
      security:
        - BearerAuth: []
      parameters:
        - name: nombre
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ciudades que coinciden con el nombre.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ciudad'
  /ciudades/{id}:
    patch:
      tags: [ciudades]
      summary: Actualizar ciudad (solo ADMIN)
      operationId: actualizarCiudad
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCiudadRequest'
      responses:
        '200':
          description: Ciudad actualizada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ciudad'
    delete:
      tags: [ciudades]
      summary: Eliminar ciudad (solo ADMIN)
      operationId: eliminarCiudad
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Ciudad eliminada.

  # ── ADMIN ─────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido en `POST /auth/login`.
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
      description: Identificador numérico del recurso.
...
  schemas:
...
    Banda:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        nombre:
          type: string
        estiloMusical:
          type: string
        historia:
          type: string
          nullable: true
        numeroComponentes:
          type: integer
          nullable: true
        imagenLogo:
          type: string
          nullable: true
        verificada:
          type: boolean
        ciudad:
          $ref: '#/components/schemas/Ciudad'
        repertorio:
          type: array
          items:
            $ref: '#/components/schemas/Marcha'
        eventos:
          type: array
          items:
            $ref: '#/components/schemas/Evento'

    # ── Procesiones ───────────────────────────
...
    Evento:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        titulo:
          type: string
        fechaHora:
          type: string
          format: date-time
        lugar:
          type: string
        descripcion:
          type: string
          nullable: true
        tipo:
          type: string
          nullable: true
        bandaId:
          type: integer
...
    CreateCiudadRequest:
      type: object
      required: [nombre, provincia, pais]
      properties:
        nombre:
          type: string
          example: "Sevilla"
        provincia:
          type: string
          example: "Sevilla"
        pais:
          type: string
          default: "España"
    UpdateCiudadRequest:
      type: object
      properties:
        nombre:
          type: string
        provincia:
          type: string
        pais:
          type: string
    Ciudad:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        nombre:
          type: string
        provincia:
          type: string
        pais:
          type: string
...
    Marcha:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        titulo:
          type: string
        compositor:
          type: string
          nullable: true

    # ── Admin ─────────────────────────────────
...
security:
  - BearerAuth: []